Update Stacks.js
A stack is an ordered collection of items that follows the LIFO (Last In First Out) principle.
The addition of new items or the removal of existing items takes place at the same end. The
end of the stack is known as the top, and the opposite side is known as the base. The newest
elements are near the top, and the oldest elements are near the base